pyproject.toml
[project]
name = "sturdy-barnacle"
version = "0.1.0"
description = "TBD"
readme = "README.md"
requires-python = ">=3.11,<3.14"
license = "MIT"
authors = [
  { name = "Tejas T A" },
]
dependencies = [
  "numpy<2.4",
  "pandas",
  "scikit-learn",
  "torch",
  "torchvision",
  "Pillow",
  "tqdm",
  "python-dotenv",
  "jupyter",
  "pinecone",
  "python-docx",
  "langchain-text-splitters",
  "langchain-openai",
  "ipykernel",
  "ipywidgets",
  "gradio",
  "packaging",
  "xgboost>=3.2.0",
  "matplotlib",
  "seaborn",
  "plotly",
  "dash",
  "optuna>=4.7.0",
  "shap>=0.50.0",
  "numba>=0.60.0",
  "llvmlite>=0.43.0",
  "mcp>=1.10.0",
  "langchain-mcp-adapters>=0.1.7",
  "langgraph>=0.1.0",
  "pyreadline3",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "ruff",
  "black",
  "mypy",
]

[project.urls]
Repository = "https://github.com/Tejas-TA/sturdy-barnacle"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[tool.ruff]
line-length = 88
target-version = "py311"
select = ["E", "F", "I", "UP"]
ignore = []

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.hatch.build.targets.wheel]
packages = ["src/deep_agent"]
